reworked outputs of input module to process GFFs and generate transcript, exon, and intron dataframes
