+++ .gitignore
@@ -7,3 +7,4 @@ __pycache__/
 data/
 tmp/
 todo.md
+classifier_todo.md
